requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
idna==3.4
Pillow==10.0.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
sentry-sdk==1.29.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
